kernel/scheduler: Add missing initializations to ThreadData::_InitBase().
fQuantumStart and fLastInterruptTime were not set to 0 here, so they
would default to the "malloc-cleared" data and then always overflow
the first time the interrupt time was tracked. I can't find any
reason that was supposed to be the behavior, so just set them to 0.

Also reorder the field initializations to be the same as the class
definition, which should allow some store merging optimizations.

Spotted by KUBSAN.