ProcessController: fix bars height computation
The previous code used a fixed margin, which would lead to too small
bars with small font sizes (the bar content would not even be visible).

Instead, use the font height as the base for the bar height, which
allows us to use the same code everywhere, no matter the menu item
height (which can vary depending on if there is an icon, etc)

Fixes #15391.