libroot: Make glibc use our math.h.
Only a few definitions (and some "hacks", to force glibc to
use __builtin_* where possible) that are not in our math.h
remain. This cuts out a lot more of the "bits" headers.

No "functional" change intended (but should help
fix the build on arches where we do not include
__fpclassify, etc. anymore.)