app_server: fix fallback for fonts where the variants match.
This does not produce a satisfactory fallback if i.e "extrabold"
is selected since the CJK font does not have that, probably need to
figure out a way to change the font style so that it renders correctly
(render cjk as bold when display is at extrabold or so).

Fixes #15751.

Signed-off-by: Augustin Cavalier <waddlesplash@gmail.com>