Appearance Prefs: Remember saved Decorator and ControlLook setting
... and make the Defaults and Revert buttons revert back to the
default/saved decorator and control look.

Add private _SetControlLook() method to set the control look which is
the equivalent of _SetDecor() but for control looks.

The person who implemented the ControlLook menu field made it so that
it would truncate the trailing "ControlLook" from the control look name,
do the same for Decorators so that BeDecorator is listed as simply "Be"
in the menu field. This is an alteration to the menu item label only and
does not affect the name of the decorator or control look.

Add private _DecorLabel() and _ControlLookLabel() methods to build
the truncated Decorator and ControlLook menu item labels respectively.

Set the current and saved decorator and control look in the constructor
and get rid of some private setter methods _AdoptToCurrentDecor() and
_AdoptInterfaceToCurrentDecor() that are no longer needed. These actions
are now done exclusively in the constructor and _SetDecor() respectively.

Remove unused knob style message constants (for now...)

Update copyright year to 2020

Change-Id: I0444a6198116d7699a84487cf97e9d826b14d1cb
Reviewed-on: https://review.haiku-os.org/c/haiku/+/2315
Reviewed-by: Adrien Destugues <pulkomandy@gmail.com>