Fix missing space/add newline in alert text
The text defined as ALERT_MSG_LOGS_USER_GUIDE is added to currently
two alert texts. If those lack a space at the end - one of them did -
there'll be a space missing after the full stop.

Adding \n to the start of ALERT_MSG_LOGS_USER_GUIDE solves this and,
as it's relatively much text, it reads better in its own paragraph.